add bullmq and config, add first queue and worker, refactor task controller to use our new bullmq jobs to run the email sending logic in the background, much more reliable - TODO: add BullMq bull-board
This realisitically in an industry standard project - These should be
decoupled, having our workers in one dir and file and then queues in
another dir and file. The only reason im keeping them coupled is due to
having to use "node <path_to_worker>" in our package.json and for
development and learning purposes i want to keep it simpler.